:recycle: refactor : ContentType.java, DeleteStatus.java 리팩토링
