+++ src/main/java/nextstep/qna/domain/Answer.java
@@ -4,6 +4,7 @@ import nextstep.global.domain.BaseTimeDomain;
 import nextstep.qna.CannotDeleteException;
 import nextstep.qna.NotFoundException;
 import nextstep.qna.UnAuthorizedException;
+import nextstep.qna.domain.enums.ContentType;
 import nextstep.qna.domain.enums.DeleteStatus;
 import nextstep.qna.domain.vo.AnswerDetail;
 import nextstep.users.domain.NsUser;
