+++ src/main/java/nextstep/qna/domain/enums/DeleteStatus.java
@@ -1,7 +1,5 @@
 package nextstep.qna.domain.enums;
 
 public enum DeleteStatus {
-    YES,
-    NO
-    ;
+    YES, NO;
 }
